+++ blocklets/ai-runtime/api/src/routes/ai.ts
@@ -33,6 +33,8 @@ import { Router } from 'express';
 import Joi from 'joi';
 import { omitBy, pick } from 'lodash';
 
+import { createLogs, mergeLogs } from './ai-log';
+
 const router = Router();
 
 const callInputSchema = Joi.object<{
@@ -274,8 +276,6 @@ router.post('/call', middlewares.session({ componentCall: true }), compression()
 
   const outputs: History['outputs'] = {};
 
-  const executingLogs: { [key: string]: NonNullable<History['steps']>[number] } = {};
-
   const taskId = nextTaskId();
 
   const sessionId =
@@ -297,17 +297,22 @@ router.post('/call', middlewares.session({ componentCall: true }), compression()
     status: 'generating',
     blockletDid,
     projectRef,
+    runType: (req.query.runType as 'cron' | 'webhook' | 'agent') || 'agent',
   });
+
+  const { logs, messageId, responseId } = createLogs({ type: 'debug', gitRef: req.body.ref, parameters: input.inputs });
+  const emitLogState = mergeLogs({ logs, messageId, responseId });
 
   const emit: RunAssistantCallback = (response) => {
     // skip debug message in production
+    const data = { ...response, messageId: history.id, sessionId };
+    emitLogState(data);
+
     if (!input.debug && response.type !== AssistantResponseType.ERROR) {
       if (response.type !== AssistantResponseType.CHUNK || (mainTaskId && response.taskId !== mainTaskId)) {
         return;
       }
     }
-
-    const data = { ...response, messageId: history.id, sessionId };
 
     if (data.type === AssistantResponseType.CHUNK || data.type === AssistantResponseType.INPUT) {
       if (data.type === AssistantResponseType.CHUNK) {
@@ -321,25 +326,6 @@ router.post('/call', middlewares.session({ componentCall: true }), compression()
             outputs.objects ??= [];
             outputs.objects.push(data.delta.object);
           }
-        }
-      }
-
-      executingLogs[data.taskId] ??= {
-        id: data.taskId,
-        agentId: data.assistantId,
-        startTime: new Date().toISOString(),
-        endTime: new Date().toISOString(),
-      };
-
-      const log = executingLogs[data.taskId]!;
-
-      // 最后一次 callback 时间作为 task 结束时间
-      log.endTime = new Date().toISOString();
-
-      if (data.type === AssistantResponseType.CHUNK) {
-        if (data.delta.object) {
-          log.objects ??= [];
-          log.objects.push(data.delta.object);
         }
       }
     } else if (data.type === AssistantResponseType.ERROR) {
@@ -358,7 +344,9 @@ router.post('/call', middlewares.session({ componentCall: true }), compression()
     res.flush();
   };
 
-  if (stream) emit({ type: AssistantResponseType.CHUNK, taskId, assistantId: agent.id, delta: {} });
+  if (stream || (history.runType && ['cron', 'webhook'].includes(history.runType))) {
+    emit({ type: AssistantResponseType.CHUNK, taskId, assistantId: agent.id, delta: {} });
+  }
 
   try {
     if (!bypassRequestLimit) {
@@ -499,9 +487,14 @@ router.post('/call', middlewares.session({ componentCall: true }), compression()
     if (typeof fetchErrorMsg !== 'string') fetchErrorMsg = fetchErrorMsg?.message;
 
     error = { ...pick(e, 'type', 'timestamp'), message: fetchErrorMsg || e.message };
+
     if (stream) {
       emit({ type: AssistantResponseType.ERROR, error });
     } else {
+      if (history.runType && ['cron', 'webhook'].includes(history.runType)) {
+        emit({ type: AssistantResponseType.ERROR, error });
+      }
+
       res.status(500).json({ error });
     }
     res.end();
@@ -513,7 +506,7 @@ router.post('/call', middlewares.session({ componentCall: true }), compression()
     error,
     outputs,
     status: 'done',
-    steps: Object.values(executingLogs),
+    logs,
     usage:
       usage.promptTokens || usage.completionTokens
         ? { ...usage, totalTokens: usage.promptTokens + usage.completionTokens }