+++ blocklets/ai-studio/src/pages/project/project-page.tsx
@@ -265,7 +265,15 @@ function ProjectPageView() {
                 variant="scrollable"
                 scrollButtons={false}
                 value={currentTab}
-                onChange={(_, tab) => setCurrentTab(tab)}
+                onChange={(_, tab) => {
+                  if (tab === 'log') {
+                    navigate(joinURL(`../../../${projectId}/logs/main?agentId=${fileId}`), {
+                      state: { fromApp: true },
+                    });
+                  } else {
+                    setCurrentTab(tab);
+                  }
+                }}
                 TabIndicatorProps={{ children: <Box component="span" /> }}
                 sx={{
                   ml: -1,
@@ -292,6 +300,7 @@ function ProjectPageView() {
                 <Tab value="debug" label={t('debug')} data-testid="debug-view-debug" />
                 <Tab value="test" label={t('test')} data-testid="debug-view-tests" />
                 <Tab value="discuss" label={t('discuss')} data-testid="debug-view-collaboration" />
+                <Tab value="log" label={t('log')} data-testid="debug-view-log" />
               </Tabs>
 
               <Box flex={1} />